add events module
